# rubocop: reduce hash literals in loops for performance