<!--
Thank you for contributing to Fluentd!
Your commits need to follow DCO: https://probot.github.io/apps/dco/
And please provide the following information to help us make the most of your pull request:
-->

**Which issue(s) this PR fixes**: 
Fixes #

**What this PR does / why we need it**: 

This is cosmetic change, it does not change behavior at all.
The following rubocop configuration detected it:

```
  Performance/CollectionLiteralInLoop:
    Enabled: true
```

hash literals in loops should not be used.

NOTE: hash literals in loops exists in test/*, but it is overkill to
fix them.

Benchmark result:

```
ruby 3.2.8 (2025-03-26 revision 13f495dc2c) +YJIT [x86_64-linux]
Warming up --------------------------------------
            in loops   550.000 i/100ms
        out of loops     1.464k i/100ms
Calculating -------------------------------------
            in loops      5.452k (± 1.5%) i/s  (183.43 μs/i) -     27.500k in   5.045506s
        out of loops     14.592k (± 1.0%) i/s   (68.53 μs/i) -     73.200k in   5.016839s

Comparison:
        out of loops:    14592.4 i/s
            in loops:     5451.7 i/s - 2.68x  slower
```

Appendix: benchmark script

```ruby
require 'bundler/inline'
gemfile do
  source 'https://rubygems.org'
  gem 'benchmark-ips'
  gem 'benchmark-memory'
end

Benchmark.ips do |x|
  x.report("in loops") {
    1000.times do |n|
      ['source', 'match', 'filter', 'label'].include?("dummy")
    end
  }
  x.report("out of loops") {
    supported_directives = ['source', 'match', 'filter', 'label']
    1000.times do |n|
      supported_directives.include?("dummy")
    end
  }
  x.compare!
end
```


**Docs Changes**:

N/A

**Release Note**: 

N/A